type declaration fixed
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@716 8aca7d54-2c30-db11-9de9-000461428c89

+++ simsys_d.c
@@ -449,7 +449,7 @@ unsigned long dr_time(void)
 }
 
 
-void dr_sleep(unsigned long usec)
+void dr_sleep(uint32 usec)
 {
 	if (usec > 1023) {
 		// schlaeft meist etwas zu kurz,
